@@ -2,75 +2,85 @@
   <fieldset class="form-item checkbox-filter">
     <legend>{{ legend }} <span v-if="!required" class="label-optional">(Optioneel)</span></legend>
 
-    <div :id="id"
-         :class="`modal modal--fixed-height checkbox-filter__modal${modalOpen ? ' visible' : ''}`"
-         tabindex="-1">
-      <div class="modal-inner">
-        <div class="modal-header">
-          <button :data-target="id"
-                  type="button"
-                  class="button icon-cross modal-close checkbox-filter__close"
-                  @click="close">
-            <span>Sluiten</span><i class="icon-close" aria-hidden="true"/>
-          </button>
-        </div>
-        <div class="modal-content">
-          <h3>{{ legend }} <span v-if="!required" class="label-optional">(Optioneel)</span></h3>
-          <div class="form-item">
-            <label :for="`checkboxes__filter_id_${legend}`">Filter onderstaande lijst</label>
-            <input :id="`checkboxes__filter_id_${legend}`" type="search"
-                   class="checkbox-filter__filter">
-            <div class="checkbox-filter__selected">
-              <span v-for="(value, index) in selectedItems" :key="`selected-${index}`"
-                    :data-value="value"
-                    class="tag filter">
-                {{ value }}
-                <button type="button" @click="removeTag(value)">
-                  <span class="visually-hidden">Verwijder tag</span>
-                </button>
-              </span>
+    <div class="form-columns">
+      <div>
+        <div :id="id"
+             :class="`modal modal--fixed-height checkbox-filter__modal${modalOpen ? ' visible' : ''}`"
+             :data-scrollable="'#' + id + '_scrollable'"
+             tabindex="-1">
+          <div class="modal-inner">
+            <div class="modal-header">
+              <button :data-target="id"
+                      type="button"
+                      class="button icon-cross modal-close checkbox-filter__close"
+                      @click="close">
+                <span>Sluiten</span><i class="icon-close" aria-hidden="true" />
+              </button>
+            </div>
+            <div :id="id + '_scrollable'" class="modal-content">
+              <h3>{{ legend }} <span v-if="!required" class="label-optional">(Optioneel)</span></h3>
+              <div class="checkbox-filter__selected">
+                <span v-for="(item, index) in selectedItems" :key="`selected-${index}`"
+                      :data-value="item"
+                      class="tag filter">
+                  {{ item }}
+                  <button type="button" @click="removeTag(item)">
+                    <span class="visually-hidden">Verwijder tag</span>
+                  </button>
+                </span>
+              </div>
+              <div class="form-item">
+                <label :for="`checkboxes__filter_id_${legend}`">Filter onderstaande lijst</label>
+                <div class="form-item checkbox-filter__filter__search-wrapper">
+                  <input :id="`checkboxes__filter_id_${legend}`" type="search"
+                         class="checkbox-filter__filter">
+                  <p class="checkbox-filter__result-wrapper" aria-live="polite" aria-atomic="true">
+                    We vonden <span class="checkbox-filter__result">#</span> resultaten.
+                  </p>
+                </div>
+              </div>
+              <div class="checkbox-filter__checkboxes">
+                <div v-for="(item, index) in items" :key="`${name}-chk-${index}`" class="checkbox">
+                  <input :id="`${name}-chk-${index}`" v-model="selectedItems"
+                         :value="item"
+                         :name="name" type="checkbox"
+                         @change.prevent="updateValue">
+                  <label :for="`${name}-chk-${index}`">{{ item }}</label>
+                </div>
+              </div>
+            </div>
+            <div class="modal-actions">
+              <button :data-target="id"
+                      type="button"
+                      class="button button-primary button-small checkbox-filter__submit modal-close"
+                      @click="updateCount">
+                Bevestig selectie
+              </button>
             </div>
           </div>
-          <p class="checkbox-filter__result-wrapper">
-            <strong aria-live="polite" class="checkbox-filter__result-wrapper">
-              We vonden <span class="checkbox-filter__result">#</span> resultaten.
-            </strong>
-          </p>
-          <div class="checkbox-filter__checkboxes">
-            <div v-for="(value, index) in items" :key="`${name}-chk-${index}`" class="checkbox">
-              <input :value="value" :id="`${name}-chk-${index}`"
-                     v-model="selectedItems"
-                     :name="name" type="checkbox"
-                     @change.prevent="updateValue">
-              <label :for="`${name}-chk-${index}`">{{ value }}</label>
-            </div>
-          </div>
+
+          <div :data-target="id"
+               class="modal-overlay modal-close"
+               @click="close" />
         </div>
-        <div class="modal-actions">
-          <button :data-target="id"
-                  type="button"
-                  class="button button-primary button-small checkbox-filter__submit modal-close"
-                  @click="updateCount">Bevestig selectie</button>
-        </div>
+
+        <p v-if="selectedCount > 0" class="checkbox-filter__count-wrapper">
+          <strong>
+            <span class="checkbox-filter__count" />
+            {{ `${selectedCount} ${selectedLegend} geselecteerd` }}
+          </strong>
+        </p>
+
+        <button :data-hash="hash"
+                :aria-controls="id"
+                type="button"
+                class="button button-secondary button-small checkbox-filter__open"
+                aria-expanded="false"
+                @click="open">
+          Selecteer ...
+        </button>
       </div>
-
-      <div :data-target="id"
-           class="modal-overlay modal-close"
-           @click="close" />
     </div>
-
-    <p v-if="selectedCount > 0">
-      <strong><span class="checkbox-filter__count"/>{{ `${selectedCount} ${selected_legend} geselecteerd` }}</strong>
-    </p>
-
-    <button :data-hash="hash"
-            :aria-controls="id"
-            type="button"
-            class="button button-secondary button-small checkbox-filter__open"
-            aria-expanded="false"
-            @click="open">
-      Selecteer ...
-    </button>
   </fieldset>
 </template>
 
@@ -91,7 +101,7 @@ export default {
       type: String,
       required: true
     },
-    selected_legend: {
+    selectedLegend: {
       type: String,
       required: true
     },
@@ -111,7 +121,7 @@ export default {
       }
     }
   },
-  data() {
+  data () {
     return {
       selectedItems: this.value,
       tempItems: [],
@@ -120,21 +130,21 @@ export default {
       hash: `#${this.id}`
     }
   },
-  mounted() {
+  mounted () {
     this.updateCount()
   },
   methods: {
     /**
      * Emit the selected items.
      */
-    updateValue() {
-      this.$emit("input", this.selectedItems)
+    updateValue () {
+      this.$emit('input', this.selectedItems)
     },
     /**
      * Remove an item from selectedItems.
      * @param {String} tag
      */
-    removeTag(tag) {
+    removeTag (tag) {
       const index = this.selectedItems.indexOf(tag)
       if (index > -1) {
         this.selectedItems.splice(index, 1)
@@ -143,7 +153,7 @@ export default {
     /**
      * Close the modal and restore selectedItems.
      */
-    close() {
+    close () {
       this.selectedItems = this.tempItems
       this.updateCount()
       this.updateValue()
@@ -152,7 +162,7 @@ export default {
     /**
      * Open the modal and save selectedItems.
      */
-    open() {
+    open () {
       this.modalOpen = true
       // make a shallow copy
       this.tempItems = this.selectedItems.slice()
@@ -160,7 +170,7 @@ export default {
     /**
      * Updated the selected items count.
      */
-    updateCount() {
+    updateCount () {
       this.modalOpen = false
       this.selectedCount = this.selectedItems.length
       this.modalOpen = false
