@@ -9,7 +9,8 @@
         <a href="https://stad.gent/over-gent-en-het-stadsbestuur/over-gent/kaarten-cijfers-en-data/gent-kaart/stadsplan">Stadsplan</a>
       </li>
       <li>
-      <a href="https://persruimte.stad.gent/">Persruimte Stad Gent</a></li>
+        <a href="https://persruimte.stad.gent/">Persruimte Stad Gent</a>
+      </li>
       <li>
         <a href="https://stad.gent/over-gent-en-het-stadsbestuur/vragen-suggesties-en-meldingen/juridische-info">Juridische info</a>
       </li>
