@@ -1,19 +1,23 @@
 <template>
   <div class="gentinfo-block">
     <div class="gentinfo-block-content">
-      <h2>Gentinfo</h2>
-      <p class="">Bereikbaar van maandag t.e.m. zaterdag van 8:00 tot 19:00 uur.</p>
+      <h2>
+        <img src="~assets/styleguide/img/svg/gentinfo-logo.svg" alt="Gentinfo" class="gentinfo-logo">
+      </h2>
+      <p>
+        Bereikbaar van maandag t.e.m. zaterdag van 8:00 tot 19:00 uur.
+      </p>
       <ul class="icon-list">
         <li>
-          <i class="icon-phone" aria-hidden="true"/>
+          <i class="icon-phone" aria-hidden="true" />
           <a href="tel:09 210 10 10" alt="Gentinfo tel link">09 210 10 10</a>
         </li>
         <li>
-          <i class="icon-envelope" aria-hidden="true"/>
+          <i class="icon-envelope" aria-hidden="true" />
           <a href="mailto:gentinfo@stad.gent" alt="Gentinfo mail link">Gentinfo@stad.gent</a>
         </li>
         <li>
-          <i class="icon-document" aria-hidden="true"/>
+          <i class="icon-document" aria-hidden="true" />
           <a href="https://stad.gent/contact" alt="Contact form">Contactformulier</a>
         </li>
       </ul>
