@@ -1,11 +1,14 @@
 <template>
   <nav v-if="total > 1" class="pager" aria-labelledby="pagination_1">
-    <h2 id="pagination_1" class="visually-hidden">Pagination</h2>
+    <h2 id="pagination_1" class="visually-hidden">
+      Pagination
+    </h2>
     <ul class="pager__items">
       <li v-if="1 !== +active" class="previous">
         <nuxt-link :to="{path: `${this.$route.path}#result`, query: Object.assign({}, $route.query,{page: +active-1})}"
                    class="previous"
-                   @click.native="blur">
+                   @click.native="blur"
+        >
           vorige
           <span class="visually-hidden">pagina</span>
         </nuxt-link>
@@ -14,16 +17,20 @@
         <nuxt-link :to="{path: `${this.$route.path}#result`, query: Object.assign({}, $route.query,{page: 1})}"
                    :title="`Ga naar pagina ${1}`"
                    :class="{active: 1 === +active}"
-                   @click.native="blur">
+                   @click.native="blur"
+        >
           <span class="visually-hidden">Pagina</span>
           {{ 1 }}
         </nuxt-link>
       </li>
-      <li v-if="active > 3">...</li>
+      <li v-if="active > 3">
+        ...
+      </li>
       <li v-if="active - 1 > 1 && +active-1 !== total">
         <nuxt-link :to="{path: `${this.$route.path}#result`, query: Object.assign({}, $route.query,{page: +active-1})}"
                    :title="`Ga naar pagina ${active-1}`"
-                   @click.native="blur">
+                   @click.native="blur"
+        >
           <span class="visually-hidden">Pagina</span>
           {{ +active - 1 }}
         </nuxt-link>
@@ -32,7 +39,8 @@
         <nuxt-link :to="{path: `${this.$route.path}#result`, query: Object.assign({}, $route.query,{page: +active})}"
                    :title="`Ga naar pagina ${active}`"
                    class="active"
-                   @click.native="blur">
+                   @click.native="blur"
+        >
           <span class="visually-hidden">Pagina</span>
           {{ +active }}
         </nuxt-link>
@@ -40,17 +48,21 @@
       <li v-if="active+1 < total">
         <nuxt-link :to="{path: `${this.$route.path}#result`, query: Object.assign({}, $route.query,{page: +active+1})}"
                    :title="`Ga naar pagina ${ +active+1}`"
-                   @click.native="blur">
+                   @click.native="blur"
+        >
           <span class="visually-hidden">Pagina</span>
           {{ +active + 1 }}
         </nuxt-link>
       </li>
-      <li v-if="total - active > 2">...</li>
+      <li v-if="total - active > 2">
+        ...
+      </li>
       <li>
         <nuxt-link :to="{path: `${this.$route.path}#result`, query: Object.assign({}, $route.query, {page: total})}"
                    :title="`Ga naar pagina ${total}`"
                    :class="{active: total === +active}"
-                   @click.native="blur">
+                   @click.native="blur"
+        >
           <span class="visually-hidden">Pagina</span>
           {{ total }}
         </nuxt-link>
@@ -58,7 +70,8 @@
       <li v-if="total !== +active" class="next">
         <nuxt-link :to="{path: `${this.$route.path}#result`, query: Object.assign({}, $route.query, {page: +active+1})}"
                    class="next"
-                   @click.native="blur">
+                   @click.native="blur"
+        >
           volgende
           <span class="visually-hidden">pagina</span>
         </nuxt-link>
@@ -90,7 +103,7 @@ export default {
      *
      * @param {ClickEvent} e
      */
-    blur: e => {
+    blur: (e) => {
       e.target.blur()
     }
   }
