@@ -1,6 +1,8 @@
 <template>
   <div class="db-block-partnership">
-    <p class="">In samenwerking met Digipolis</p>
-    <i aria-hidden="true" class="icon-digipolis"/>
+    <p>
+      In samenwerking met Digipolis
+    </p>
+    <i aria-hidden="true" class="icon-digipolis" />
   </div>
 </template>
