@@ -8,7 +8,9 @@
         <div class="modal-header">
           <button type="button"
                   class="button close icon-cross modal-close"
-                  @click="closeModal">Sluiten</button>
+                  @click="closeModal">
+            Sluiten
+          </button>
         </div>
         <div class="modal-content">
           <h2>Zoek verwerking</h2>
@@ -17,84 +19,91 @@
               <label for="name">Naam <span class="label-optional">(Optioneel)</span></label>
               <input id="name" v-model="filter.name" type="text" name="name">
             </div>
-            <checkbox_with_filter v-if="processors.length"
+            <checkbox-with-filter v-if="processors.length"
                                   :id="'services'"
+                                  v-model="filter['processor[]']"
                                   :items="processors"
                                   :legend="'Verwerkende dienst'"
-                                  :selected_legend="'dienst(en)'"
-                                  :name="'processor[]'"
-                                  v-model="filter['processor[]']"/>
-            <checkbox_with_filter v-if="personalData.length"
+                                  :selected-legend="'dienst(en)'"
+                                  :name="'processor[]'" />
+            <checkbox-with-filter v-if="personalData.length"
                                   :id="'data'"
+                                  v-model="filter['personalData[]']"
                                   :items="personalData"
                                   :legend="'Welke gegevens?'"
-                                  :selected_legend="'gegeven(s)'"
-                                  :name="'personalData[]'"
-                                  v-model="filter['personalData[]']"/>
-            <checkbox_with_filter v-if="grantees.length"
+                                  :selected-legend="'gegeven(s)'"
+                                  :name="'personalData[]'" />
+            <checkbox-with-filter v-if="grantees.length"
                                   :id="'recipient'"
+                                  v-model="filter['grantees[]']"
                                   :items="grantees"
                                   :legend="'Ontvanger'"
-                                  :selected_legend="'ontvanger(s)'"
-                                  :name="'grantees[]'"
-                                  v-model="filter['grantees[]']"/>
-            <checkbox_with_filter v-if="types.length"
+                                  :selected-legend="'ontvanger(s)'"
+                                  :name="'grantees[]'" />
+            <checkbox-with-filter v-if="types.length"
                                   :id="'category'"
+                                  v-model="filter['types[]']"
                                   :items="types"
                                   :legend="'Categorie'"
-                                  :selected_legend="'categorie(ën)'"
-                                  :name="'types[]'"
-                                  v-model="filter['types[]']"/>
+                                  :selected-legend="'categorie(ën)'"
+                                  :name="'types[]'" />
 
             <fieldset v-if="formalFrameworks.length" class="form-item">
               <legend>Rechtmatigheid</legend>
-              <div v-for="(formalFramework, index) in formalFrameworks" :key="index" class="checkbox" >
+              <div v-for="(formalFramework, index) in formalFrameworks" :key="index" class="checkbox">
                 <input :id="`formalFrameworks-chk-${index}`"
+                       v-model="filter['formalFrameworks[]']"
                        :value="formalFramework"
                        :name="'formalFrameworks[]'"
-                       v-model="filter['formalFrameworks[]']"
                        type="checkbox"
                        class="checkbox">
                 <label :for="`formalFrameworks-chk-${index}`">{{ formalFramework }}</label>
               </div>
             </fieldset>
-            <button type="submit" class="button button-primary filter__submit modal-close" @click="closeModal">Zoek</button>
+            <button type="submit" class="button button-primary filter__submit modal-close" @click="closeModal">
+              Zoek
+            </button>
           </form>
         </div>
       </div>
     </section>
     <section id="result" class="content result-section">
-      <selectedfilters :allowed-filters="allowedFilters"/>
-      <button type="button"
-              class="button button-primary icon-filter result__show-filters modal-trigger"
-              aria-expanded="false"
-              aria-controls="filter"
-              @click="openModal">Toon filters</button>
-      <h2 :class="{'visually-hidden': selectedFilters.length === 0}">We vonden {{ total }} {{ total === 1 ? 'resultaat' : 'resultaten' }}</h2>
+      <selectedfilters :allowed-filters="allowedFilters" />
+      <div class="filter__result-count">
+        <h2 :class="{'visually-hidden': selectedFilters.length === 0}">
+          We vonden {{ total }} {{ total === 1 ? 'resultaat' : 'resultaten' }}
+        </h2>
+        <button type="button"
+                class="button button-secondary icon-filter result__show-filters modal-trigger"
+                aria-expanded="false"
+                aria-controls="filter"
+                @click="openModal">
+          Toon filters
+        </button>
+      </div>
       <ul class="filter__results">
         <teaser v-for="(item, index) in paginatedItems"
                 :key="index"
                 :item="item"
-                :index="index"/>
+                :index="index" />
       </ul>
       <pagination
         :total="totalPages"
-        :active="currentPage"/>
+        :active="currentPage" />
     </section>
   </div>
 </template>
 
 <script>
-import teaser from "~/components/molecules/teaser"
-import pagination from "~/components/molecules/pagination"
-import selectedfilters from "~/components/molecules/selectedfilters"
-import checkbox_with_filter from "~/components/molecules/checkbox-with-filter"
-import introductietekst from "~/components/introductietekst"
+import teaser from '~/components/molecules/teaser'
+import pagination from '~/components/molecules/pagination'
+import selectedfilters from '~/components/molecules/selectedfilters'
+import checkboxWithFilter from '~/components/molecules/checkbox-with-filter'
 
 const sortFunction = (a, b) => {
   // omit non-word characters
-  a = a.replace(/\W/g, "").toUpperCase()
-  b = b.replace(/\W/g, "").toUpperCase()
+  a = a.replace(/\W/g, '').toUpperCase()
+  b = b.replace(/\W/g, '').toUpperCase()
 
   if (a > b) {
     return 1
@@ -110,8 +119,7 @@ export default {
     teaser,
     pagination,
     selectedfilters,
-    checkbox_with_filter,
-    introductietekst
+    checkboxWithFilter
   },
   props: {
     items: {
@@ -119,31 +127,31 @@ export default {
       required: true
     }
   },
-  data() {
+  data () {
     return {
       itemsPerPage: 10,
       allowedFilters: [
-        "name",
-        "grantees[]",
-        "personalData[]",
-        "types[]",
-        "formalFrameworks[]",
-        "processor[]"
+        'name',
+        'grantees[]',
+        'personalData[]',
+        'types[]',
+        'formalFrameworks[]',
+        'processor[]'
       ],
       filter: {
         name: this.$route.query.name,
-        "personalData[]": this.parseQueryArray("personalData[]") || [],
-        "grantees[]": this.parseQueryArray("grantees[]") || [],
-        "types[]": this.parseQueryArray("types[]") || [],
-        "formalFrameworks[]": this.parseQueryArray("formalFrameworks[]") || [],
-        "processor[]": this.parseQueryArray("processor[]") || []
+        'personalData[]': this.parseQueryArray('personalData[]') || [],
+        'grantees[]': this.parseQueryArray('grantees[]') || [],
+        'types[]': this.parseQueryArray('types[]') || [],
+        'formalFrameworks[]': this.parseQueryArray('formalFrameworks[]') || [],
+        'processor[]': this.parseQueryArray('processor[]') || []
       },
       filterHidden: false,
       modalOpen: false
     }
   },
   computed: {
-    processors() {
+    processors () {
       return this.items
         .reduce((result, item) => {
           if (
@@ -157,11 +165,11 @@ export default {
         }, [])
         .sort(sortFunction)
     },
-    personalData() {
+    personalData () {
       return this.items
         .reduce((result, item) => {
           if (item.personalData && item.personalData.value) {
-            for (let i = item.personalData.value.length; i--; ) {
+            for (let i = item.personalData.value.length; i--;) {
               if (!result.includes(item.personalData.value[i])) {
                 result.push(item.personalData.value[i])
               }
@@ -171,11 +179,11 @@ export default {
         }, [])
         .sort(sortFunction)
     },
-    grantees() {
+    grantees () {
       return this.items
         .reduce((result, item) => {
           if (item.grantees && item.grantees.value) {
-            for (let i = item.grantees.value.length; i--; ) {
+            for (let i = item.grantees.value.length; i--;) {
               if (!result.includes(item.grantees.value[i])) {
                 result.push(item.grantees.value[i])
               }
@@ -185,7 +193,7 @@ export default {
         }, [])
         .sort(sortFunction)
     },
-    types() {
+    types () {
       return this.items
         .reduce((result, item) => {
           if (
@@ -199,7 +207,7 @@ export default {
         }, [])
         .sort(sortFunction)
     },
-    formalFrameworks() {
+    formalFrameworks () {
       return this.items
         .reduce((result, item) => {
           if (
@@ -213,28 +221,28 @@ export default {
         }, [])
         .sort(sortFunction)
     },
-    filteredItems() {
-      const processors = this.parseQueryArray("processor[]")
+    filteredItems () {
+      const processors = this.parseQueryArray('processor[]')
       const processorRegex = processors
-        ? new RegExp(processors.join("|"), "i")
+        ? new RegExp(processors.join('|'), 'i')
         : null
 
-      const data = this.parseQueryArray("personalData[]")
-      const dataRegex = data ? new RegExp(data.join("|"), "i") : null
+      const data = this.parseQueryArray('personalData[]')
+      const dataRegex = data ? new RegExp(data.join('|'), 'i') : null
 
-      const grantees = this.parseQueryArray("grantees[]")
-      const granteeRegex = grantees ? new RegExp(grantees.join("|"), "i") : null
+      const grantees = this.parseQueryArray('grantees[]')
+      const granteeRegex = grantees ? new RegExp(grantees.join('|'), 'i') : null
 
-      const types = this.parseQueryArray("types[]")
-      const typesRegex = types ? new RegExp(types.join("|"), "i") : null
+      const types = this.parseQueryArray('types[]')
+      const typesRegex = types ? new RegExp(types.join('|'), 'i') : null
 
-      const formalFrameworks = this.parseQueryArray("formalFrameworks[]")
+      const formalFrameworks = this.parseQueryArray('formalFrameworks[]')
       const formalFrameworksRegex = formalFrameworks
-        ? new RegExp(formalFrameworks.join("|"), "i")
+        ? new RegExp(formalFrameworks.join('|'), 'i')
         : null
 
       return this.items
-        .filter(item => {
+        .filter((item) => {
           /* Check each filter and
             ** return true if all checks are valid
             */
@@ -296,8 +304,8 @@ export default {
           try {
             // sort by name
             // omit non-word characters
-            const nameA = a.name.value.replace(/\W/g, "").toUpperCase()
-            const nameB = b.name.value.replace(/\W/g, "").toUpperCase()
+            const nameA = a.name.value.replace(/\W/g, '').toUpperCase()
+            const nameB = b.name.value.replace(/\W/g, '').toUpperCase()
 
             if (nameA > nameB) {
               return 1
@@ -309,10 +317,10 @@ export default {
             // sort by processor
             // omit non-word characters
             const processorA = a.processor.value
-              .replace(/\W/g, "")
+              .replace(/\W/g, '')
               .toUpperCase()
             const processorB = b.processor.value
-              .replace(/\W/g, "")
+              .replace(/\W/g, '')
               .toUpperCase()
             if (processorA > processorB) {
               return 1
@@ -326,17 +334,17 @@ export default {
           return 0
         })
     },
-    total() {
+    total () {
       return this.filteredItems.length
     },
-    totalPages() {
+    totalPages () {
       return Math.ceil(this.total / this.itemsPerPage)
     },
-    paginatedItems() {
+    paginatedItems () {
       const index = this.currentPage * this.itemsPerPage - this.itemsPerPage
       return this.filteredItems.slice(index, index + this.itemsPerPage)
     },
-    currentPage() {
+    currentPage () {
       const queryPage = this.$route.query.page || 1
       if (queryPage <= 0 || isNaN(queryPage)) {
         return 1
@@ -346,7 +354,7 @@ export default {
       }
       return +queryPage
     },
-    selectedFilters() {
+    selectedFilters () {
       return Object.keys(this.$route.query).reduce((result, key) => {
         if (
           this.allowedFilters.includes(key) &&
@@ -362,7 +370,7 @@ export default {
       }, [])
     }
   },
-  mounted() {
+  mounted () {
     this.filterHidden = window.innerWidth > 768
   },
   methods: {
@@ -372,7 +380,7 @@ export default {
      * @param {String} key
      * @returns {Array|null}
      */
-    parseQueryArray(key) {
+    parseQueryArray (key) {
       if (!this.$route.query[key]) {
         return null
       }
@@ -384,25 +392,18 @@ export default {
     /**
      * Push selected filters to the query.
      */
-    submitFilter() {
-      /*
-        Nuxt bug.
-        URL won't update if an array in the query changes.
-        Altering the route query first solves this.
-         */
-      this.$route.query.check++
-
+    submitFilter () {
       this.$router.push({
         path: `${this.$route.path}#result`,
         // Override existing query, including pagination
         query: this.filter
       })
     },
-    openModal() {
+    openModal () {
       this.filterHidden = true
       this.modalOpen = true
     },
-    closeModal() {
+    closeModal () {
       this.filterHidden = false
       this.modalOpen = false
     }
