@@ -8,8 +8,8 @@
 export default ({ store, route }) => {
   const meta = route.meta[0]
   if (meta && meta.breadcrumbs) {
-    store.commit("SET_BREADCRUMBS", meta.breadcrumbs)
+    store.commit('SET_BREADCRUMBS', meta.breadcrumbs)
   } else {
-    store.commit("SET_BREADCRUMBS", [])
+    store.commit('SET_BREADCRUMBS', [])
   }
 }
